app/requirements.txt
aiohttp==3.9.5
aiolimiter==1.1.0
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
attrs==23.2.0
beautifulsoup4==4.12.3
boto3==1.34.106
botocore==1.34.106
bs4==0.0.2
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
distro==1.9.0
dnspython==2.6.1
email_validator==2.1.1
environs==9.5.0
fastapi==0.111.0
fastapi-cli==0.0.3
frozenlist==1.4.1
groq==0.6.0
grpcio==1.60.0
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.7
Jinja2==3.1.4
jmespath==1.0.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.2
mdurl==0.1.2
milvus-lite==2.4.3
multidict==6.0.5
numpy==1.26.4
orjson==3.10.3
packaging==24.0
pandas==2.2.2
protobuf==5.26.1
psycopg2-binary==2.9.9
pydantic==2.7.1
pydantic_core==2.18.2
Pygments==2.18.0
pymilvus==2.4.2
pymongo==4.7.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
rich==13.7.1
s3transfer==0.10.1
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
starlette==0.37.2
tenacity==8.3.0
typer==0.12.3
typing_extensions==4.11.0
tzdata==2024.1
ujson==5.9.0
urllib3==2.2.1
uvicorn==0.29.0
uvloop==0.19.0
voyageai==0.2.2
watchfiles==0.21.0
websockets==12.0
yarl==1.9.4
